+++ src/kits/network/libsocket/socket.c
@@ -128,7 +128,7 @@ _EXPORT int accept(int sock, struct sockaddr *addr, int *addrlen)
 
 
 _EXPORT ssize_t recvfrom(int sock, void *buffer, size_t buflen, int flags,
-             struct sockaddr *addr, size_t *addrlen)
+             struct sockaddr *addr, int *addrlen)
 {
 	struct msghdr mh;
 	struct iovec iov;
@@ -156,7 +156,7 @@ _EXPORT ssize_t recvfrom(int sock, void *buffer, size_t buflen, int flags,
 }
 
 _EXPORT ssize_t sendto(int sock, const void *buffer, size_t buflen, int flags,
-           const struct sockaddr *addr, size_t addrlen)
+           const struct sockaddr *addr, int addrlen)
 {
 	struct msghdr mh;
 	struct iovec iov;
